Move oitg-master-hoa2 to Nix flakes
Other targets (which are mostly in personal use by DPN)
to follow.

This required some changes to the upstream flake.nix. We'd like
to be able to do without that diff, but in the meantime, this is
tracked automatically by the branch specified in flake.nix and
the corresponding commit information in flake.lock.